// src/components/MovieLayout/MovieDetail/Interfaces/ICastWithCredits.tsx
// Generated by https://quicktype.io

export interface ICast {
  cast: IPersonWithMovieCredits[];
}

export interface IPersonWithMovieCredits {
  poster_path: string;
  video: boolean;
  vote_average: number;
  id: number;
  overview: string;
  release_date: string;
  adult: boolean;
  backdrop_path: string;
  vote_count: number;
  genre_ids: number[];
  title: string;
  original_language: string;
  original_title: string;
  popularity: number;
  character: string;
  credit_id: string;
  order: number;
}